added confirmation pop-up box to EnergizerDeleteButton component

+++ client/src/pages/components/EnergizerDeleteButton.js
@@ -11,31 +11,38 @@ function EnergizerDeleteButton({ energizerId, energizerName }) {
 
 	const handleDelete = async (e) => {
 		e.preventDefault();
-		// if the energizer is in the favourites alerts the user and asks to remove it from favourites
+		// if the energizer is in the favourites/local storage alerts the user and asks to remove it from favourites
 		if (findFavouriteEnergizer()) {
 			alert(
 				"Please remove the Energizer form your favourites before you delete this energizer"
 			);
 			return;
 		}
-		//fixed the alert, the alert now displays the name of the energizer to be removed
-		alert(`You are deleting energizer: ${energizerName}`);
-		try {
-			setIsDeleting(true);
-			const response = await fetch(`/api/energizers/${energizerId}`, {
-				method: "DELETE",
-			});
-			if (!response.ok) {
-				const errorText = await response.text();
-				throw new Error(errorText);
-			}
+		//confirmation box to validate whether you really want to remove the energizer, prevents accidental deletes
+		if (
+			window.confirm(
+				`Do you really want to delete energizer: ${energizerName}?`
+			)
+		) {
+			//fixed the alert, the alert now displays the name of the energizer to be removed
+			alert(`You are deleting energizer: ${energizerName}`);
+			try {
+				setIsDeleting(true);
+				const response = await fetch(`/api/energizers/${energizerId}`, {
+					method: "DELETE",
+				});
+				if (!response.ok) {
+					const errorText = await response.text();
+					throw new Error(errorText);
+				}
 
-			window.history.back();
-		} catch (error) {
-			console.error(error);
-			alert("Failed to delete energizer");
-		} finally {
-			setIsDeleting(false);
+				window.history.back();
+			} catch (error) {
+				console.error(error);
+				alert("Failed to delete energizer");
+			} finally {
+				setIsDeleting(false);
+			}
 		}
 	};
